requirements.txt
alembic==1.10.2  # You can specify the version you need or use the latestfastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
python-multipart==0.0.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==1.8.2
python-dotenv==0.19.0
aiofiles==0.7.0

# ML dependencies
--find-links https://download.pytorch.org/whl/torch_stable.html
torch>=2.1.0
transformers>=4.11.3
tokenizers>=0.10.3
openai-whisper>=20231117
spacy>=3.2.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0.tar.gz
numpy>=1.21.2
pandas>=1.3.3
scikit-learn>=0.24.2
